AI Image: Move components to own folder
