+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-image/components/usage-counter.scss
